# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.album:
    properties:
      artist:
        type: string
      id:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /albums:
    get:
      description: Responds with the list of all albums as JSON.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.album'
            type: array
      summary: Get albums array
      tags:
      - albums
    post:
      description: Takes a album JSON and store in DB. Return saved JSON.
      parameters:
      - description: Album JSON
        in: body
        name: album
        required: true
        schema:
          $ref: '#/definitions/main.album'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.album'
      summary: Store a new album
      tags:
      - albums
  /albums/{id}:
    get:
      description: Returns the album whose id value matches the id.
      parameters:
      - description: search album by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.album'
      summary: Get single album by id
      tags:
      - albums
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"